- Merging the branch with the partially implemented TPC-H benchmark. The branch was enceladus.db.cs.cmu.edu/svn/branches/shore-kits-sparcv9-solaris2.10-suncc-r175.
- It has integrated a version of DBGEN that it is used to populate the TPC-H database.
- It implements at least Q1 and Q6. But, the queries have not been debugged.

- Another change is that the SF is made double instead of int (for TPC-H SF=1 generates a 2.5GB database, therefore mostly for debugging reasons we need to have SF <1).

+++ include/dora/tm1/dora_tm1_client.h
@@ -58,7 +58,7 @@ private:
     // workload parameters
     DoraTM1Env* _tm1db;    
     int _selid;
-    int _qf;
+    double _qf;
 
 public:
 
@@ -67,7 +67,7 @@ public:
     dora_tm1_client_t(c_str tname, const int id, DoraTM1Env* env, 
                       const MeasurementType aType, const int trxid, 
                       const int numOfTrxs, 
-                      processorid_t aprsid, const int selID, const int qf)  
+                      processorid_t aprsid, const int selID, const double qf)  
 	: base_client_t(tname,id,env,aType,trxid,numOfTrxs,aprsid),
           _tm1db(env), _selid(selID), _qf(qf)
     {
