+++ include/dora/tpcb/dora_tpcb_client.h
@@ -56,7 +56,7 @@ private:
     // workload parameters
     DoraTPCBEnv* _tpcbdb;    
     int _selid;
-    int _qf;
+    double _qf;
 
 public:
 
@@ -65,7 +65,7 @@ public:
     dora_tpcb_client_t(c_str tname, const int id, DoraTPCBEnv* env, 
                       const MeasurementType aType, const int trxid, 
                       const int numOfTrxs, 
-                      processorid_t aprsid, const int selID, const int qf)  
+                      processorid_t aprsid, const int selID, const double qf)  
 	: base_client_t(tname,id,env,aType,trxid,numOfTrxs,aprsid),
           _tpcbdb(env), _selid(selID), _qf(qf)
     {
