+++ include/workload/tpcb/shore_tpcb_client.h
@@ -21,8 +21,6 @@
    RESULTING FROM THE USE OF THIS SOFTWARE.
 */
 
-/* -*- mode:C++; c-basic-offset:4 -*- */
-
 /** @file:   shore_tpcb_client.h
  *
  *  @brief:  Defines test client for the TPC-B benchmark
@@ -54,10 +52,9 @@ ENTER_NAMESPACE(tpcb);
 class baseline_tpcb_client_t : public base_client_t 
 {
 private:
-
     int _selid;
     trx_worker_t* _worker;
-    int _qf;    
+    double _qf;    
 
 public:
 
@@ -66,7 +63,7 @@ public:
     baseline_tpcb_client_t(c_str tname, const int id, ShoreTPCBEnv* env, 
                            const MeasurementType aType, const int trxid, 
                            const int numOfTrxs, 
-                           processorid_t aprsid, const int selID, const int qf);
+                           processorid_t aprsid, const int selID, const double qf);
 
     ~baseline_tpcb_client_t() { }
